cv-projects/RAG-Chatbot/requirements_production.txt
# Production Requirements for RAG Chatbot
# Core dependencies
streamlit>=1.28.0
ollama>=0.1.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
langchain>=0.0.350
PyPDF2>=3.0.1
numpy>=1.24.0
pandas>=2.0.0

# Production dependencies
gunicorn>=21.2.0
python-dotenv>=1.0.0
requests>=2.31.0
psutil>=5.9.0

# Optional: For better performance
# faiss-gpu>=1.7.4  # Uncomment if you have GPU support
